ai-model/requirements.txt
# Core ML & Data
torch>=2.0.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Vector database
pymilvus>=2.3.0

# Database
psycopg2-binary>=2.9.0

# Environment & Config
python-dotenv>=1.0.0

# Data I/O
pyarrow>=12.0.0  # For parquet support